QA: Flyttet Kolonnetittel, LogiskeVedlegg, Hovedtabellrad og Tabellrad komponenter i egne filer
