+++ src/frontend/Komponenter/Personoversikt/Dokumentoversikt/Tabellrad.tsx
@@ -1,0 +1,48 @@
+import React from 'react';
+import { Dokumentinfo } from '../../../App/typer/dokumentliste';
+import { Td } from '../../../Felles/Personopplysninger/TabellWrapper';
+import { LogiskeVedlegg } from './LogiskeVedlegg';
+import { BodyShortSmall } from '../../../Felles/Visningskomponenter/Tekster';
+import styled from 'styled-components';
+import { useToggles } from '../../../App/context/TogglesContext';
+import { skalViseLenke } from '../utils';
+
+const LenkeVenstreMargin = styled.a`
+    margin-left: 2rem;
+
+    &:visited {
+        color: purple;
+    }
+`;
+
+export const Tabellrad: React.FC<{ dokument: Dokumentinfo; erKlikketId: string }> = ({
+    dokument,
+}) => {
+    const { toggles } = useToggles();
+    return (
+        <tr>
+            <Td></Td>
+            <Td></Td>
+            <Td>
+                {skalViseLenke(dokument, toggles) ? (
+                    <>
+                        <LenkeVenstreMargin
+                            href={`/dokument/journalpost/${dokument.journalpostId}/dokument-pdf/${dokument.dokumentinfoId}`}
+                            target={'_blank'}
+                            rel={'noreferrer'}
+                        >
+                            {dokument.tittel}
+                        </LenkeVenstreMargin>
+                        <LogiskeVedlegg logiskeVedlegg={dokument.logiskeVedlegg} />
+                    </>
+                ) : (
+                    <BodyShortSmall>{dokument.tittel}</BodyShortSmall>
+                )}
+            </Td>
+            <Td></Td>
+            <Td></Td>
+            <Td></Td>
+            <Td></Td>
+        </tr>
+    );
+};
