+++ src/frontend/Komponenter/Personoversikt/utils.ts
@@ -2,6 +2,8 @@ import { BehandlingResultat, Fagsak } from '../../App/typer/fagsak';
 import { BehandlingStatus } from '../../App/typer/behandlingstatus';
 import { Behandlingstype } from '../../App/typer/behandlingstype';
 import { VedleggRequest } from './vedleggRequest';
+import { Dokumentinfo } from '../../App/typer/dokumentliste';
+import { ToggleName, Toggles } from '../../App/context/toggles';
 
 export const alleBehandlingerErFerdigstiltEllerSattPåVent = (fagsak: Fagsak) =>
     fagsak.behandlinger.every(
@@ -43,4 +45,11 @@ export const oppdaterVedleggFilter = (
         ...object,
         [key]: val,
     };
+};
+
+export const skalViseLenke = (dokument: Dokumentinfo, toggles: Toggles): boolean => {
+    return (
+        (toggles[ToggleName.dokumentoversiktLinkTilDokument] || dokument.tema === 'ENF') &&
+        dokument.harSaksbehandlerTilgang
+    );
 };
